requirements.txt
aiogram==3.4.1
aiohttp==3.9.5
aiosqlite==0.20.0
pydantic==2.7.1
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
